requirements.txt
Django==1.8
Markdown==2.6.1
dj-database-url==0.3.0
django-filter==0.9.2
djangorestframework==3.1.1
psycopg2==2.6
wsgiref==0.1.2
axelrod==0.0.8
redis==2.10.3
celery[redis]==3.1.18
python-social-auth==0.2.10
requests==2.7.0
jsonfield==1.0.3
